backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Cloud Services
google-cloud-speech==2.24.0
google-cloud-texttospeech==2.16.0
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
numpy==1.26.3

# HTTP & API
httpx==0.26.0
aiofiles==23.2.1

# CORS
fastapi-cors==0.0.6

# Validation & Utils
phonenumbers==8.13.27
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
